+++ scripts/install/nginx.sh
@@ -2,35 +2,50 @@
 
 # Vhosts
 echo "Loading Nginx vhosts..."
+
+TORAN_HOST=${TORAN_HOST:-localhost}
+
 rm -f /etc/nginx/sites-enabled/*
-if [ "${TORAN_HTTPS}" == "true" ]; then
 
-    echo "Loading HTTPS Certificates..."
+if [ ! -e "${DATA_DIRECTORY}/certs/toran-proxy.key" ] && [ ! -e "${DATA_DIRECTORY}/certs/toran-proxy.crt" ]; then
+    SSL_CERT_SUBJ="/C=SS/ST=SS/L=SelfSignedCity/O=SelfSignedOrg/CN=${TORAN_HOST}"
 
-    if [ ! -e "$DATA_DIRECTORY/certs/toran-proxy.key" ]; then
-        echo "ERROR: "
-        echo "  Please add toran-proxy.key in folder certs/"
-        exit 1
-    fi
+    echo "Generating self-signed HTTPS Certificates..."
 
-    if [ ! -e "$DATA_DIRECTORY/certs/toran-proxy.crt" ]; then
-        echo "ERROR: "
-        echo "  Please add toran-proxy.crt in folder certs/"
-        exit 1
-    fi
+    mkdir -p ${DATA_DIRECTORY}/certs
 
-    # Add certificates trusted
-    ln -s $DATA_DIRECTORY/certs /usr/local/share/ca-certificates/toran-proxy
-    update-ca-certificates
+    openssl req \
+        -x509 \
+        -nodes \
+        -days 365 \
+        -newkey rsa:2048 \
+        -keyout "${DATA_DIRECTORY}/certs/toran-proxy.key" \
+        -out "${DATA_DIRECTORY}/certs/toran-proxy.crt" \
+        -subj "${SSL_CERT_SUBJ}"
 
-    ln -s /etc/nginx/sites-available/toran-proxy-https.conf /etc/nginx/sites-enabled/toran-proxy-https.conf
+elif [ -e "${DATA_DIRECTORY}/certs/toran-proxy.key" ] && [ -e "${DATA_DIRECTORY}/certs/toran-proxy.crt" ]; then
+
+    echo "Using provided HTTPS Certificates..."
+
 else
-    ln -s /etc/nginx/sites-available/toran-proxy-http.conf /etc/nginx/sites-enabled/toran-proxy-http.conf
+
+    if [ ! -e "${DATA_DIRECTORY}/certs/toran-proxy.key" ]; then
+        echo "ERROR: "
+        echo "  File toran-proxy.key exists in folder certs/ but no toran-proxy.crt"
+        exit 1
+    else
+        echo "ERROR: "
+        echo "  File toran-proxy.crt exists in folder certs/ but no toran-proxy.key"
+        exit 1
+    fi
+
 fi
 
-# Logs
-mkdir -p $DATA_DIRECTORY/logs/nginx
+echo "Loading HTTPS Certificates..."
 
-# Loading permissions
-chown -R www-data:www-data \
-    $DATA_DIRECTORY/logs/nginx
+# Add certificates trusted
+mkdir -p /usr/local/share/ca-certificates/toran-proxy
+cp -f ${DATA_DIRECTORY}/certs/* /usr/local/share/ca-certificates/toran-proxy
+update-ca-certificates
+
+ln -s /etc/nginx/sites-available/toran-proxy.conf /etc/nginx/sites-enabled/toran-proxy.conf